finalized metric tab
everything is functional. im not fully happy with the css so ill be fixing it again at some point. i also want to go in and try to condense it so there arent so many files. 

same as with brandon, open the folder with these files in terminal. 
type "python -m http.server 8080" and then paste this link in chrome -> http://localhost:8080/metrics.html