updated asset menu and buttons
asset menu is pretty much fully functional besides the fill portion. still working on having the assets going to an asset pool and having that be visible globally. i think the register tab works with the back end (at least i added all the required code for it). Edit asset does not work with the backend as far as i know.